maint: Fix compiling errors on Linux due to last commit
* src/common.h: Add missing include.
* src/gingancl/FormatterActiveDevice.cpp: Updated accordingly.
* src/gingancl/FormatterActiveDevice.h: Likewise.
* src/gingancl/FormatterBaseDevice.cpp: Likewise.
* src/gingancl/FormatterBaseDevice.h: Likewise.
* src/gingancl/FormatterPassiveDevice.cpp: Likewise.
* src/isdbt-dataproc/dsmcc/StreamEvent.cpp: Likewise.
* src/isdbt-tsparser/ServiceInfo.cpp: Likewise.
* src/isdbt-tsparser/ServiceInfo.h: Likewise.
* src/multidevice/BroadcastDualSocketService.cpp: Likewise.
* src/multidevice/DeviceDomain.h: Likewise.
* src/multidevice/IRemoteDeviceListener.h: Likewise.
* src/multidevice/MulticastSocketService.cpp: Likewise.
* src/multidevice/RemoteDevice.h: Likewise.
* src/multidevice/RemoteEventService.cpp: Likewise.
* src/multidevice/RemoteEventService.h: Likewise.
* src/multidevice/TcpSocketService.cpp: Likewise.

+++ src/isdbt-dataproc/dsmcc/StreamEvent.cpp
@@ -31,7 +31,7 @@ BR_PUCRIO_TELEMIDIA_GINGA_CORE_DATAPROCESSING_BEGIN
 		this->eventId = ((data[2] & 0xFF) << 8) | (data[3] & 0xFF);
 
 		//4,5,6,7* reserved
-		this->timeReference = ((data[7] & 0x01) << 32) |
+		this->timeReference = (((uint64_t)(data[7] & 0x01)) << 32) |
 				((data[8] & 0xFF) << 24) | ((data[9] & 0xFF) << 16) |
 				((data[10] & 0xFF) << 8) | (data[11] & 0xFF);
 	}
